feat: api 이전 및 deleteId 삭제
